Env attributes for Cucumber v8 & forced colours
With the switch from "colors" to "chalk" included in the bump to Cucumber 8, something has broken in the way supports-color is integrated into pretty-formatter and no colours were being displayed on stdout.

Although it's a very small detail, it'd be nice to keep (and specially to keep in mind in case any other differences arise) and fortunately it can be circumvented by just setting the "colorsEnabled" format options to true. That's something that should be taken care of automatically, but anyways, before we can make a decision on how relevant it is to devote time to this, this fixes the implementation.